requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.28
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0 